+++ include/football/CMatch.h
@@ -1,32 +1,12 @@
 #pragma once
 
 #include "IJsonable.h"
-#include "traits/json.h"
+#include "football/traits/CMatch.h"
 
 #include "football/CTeam.h"
 #include "football/CStadium.h"
 
-namespace futsim
-{
-
-namespace football
-{
-class CMatch;
-}
-
-template <> struct json_traits<football::CMatch>
-{
-	//! JSON key for the class.
-	static inline constexpr std::string_view KEY = "Match";
-	//! JSON key for the \copybrief football::CMatch::mHomeTeam
-	static inline constexpr std::string_view HOME_TEAM_KEY = "Home team";
-	//! JSON key for the \copybrief football::CMatch::mAwayTeam
-	static inline constexpr std::string_view AWAY_TEAM_KEY = "Away team";
-	//! JSON key for the \copybrief football::CMatch::mReferee
-	static inline constexpr std::string_view REFEREE_KEY = "Referee";
-};
-
-namespace football
+namespace futsim::football
 {
 
 class CTeamStrategy;
@@ -96,6 +76,4 @@ private:
 	name_type mReferee;
 };
 
-} // football namespace
-
-} // futsim namespace
+} // futsim::football namespace
